requirements.txt
aaiogram==3.10.0
sqlalchemy==2.0.35
asyncpg==0.29.0
aiosqlite==0.20.0
apscheduler==3.10.4
python-dotenv==1.0.1
aiohttp==3.9.5
matplotlib==3.9.1
pandas==2.2.2
Pillow==10.4.0
gunicorn==21.2.0